+++ src/burn/engine/core.cpp
@@ -527,7 +527,7 @@ extern "C" HRESULT CoreApply(
     BOOL fSuspend = FALSE;
     BOOTSTRAPPER_APPLY_RESTART restart = BOOTSTRAPPER_APPLY_RESTART_NONE;
     BURN_CACHE_THREAD_CONTEXT cacheThreadContext = { };
-    DWORD dwNumberOfPhases = 0;
+    DWORD dwPhaseCount = 0;
 
     LogId(REPORT_STANDARD, MSG_APPLY_BEGIN);
 
@@ -539,13 +539,13 @@ extern "C" HRESULT CoreApply(
 
     if (pEngineState->plan.cCacheActions)
     {
-        ++dwNumberOfPhases;
+        ++dwPhaseCount;
     }
     if (pEngineState->plan.cExecuteActions)
     {
-        ++dwNumberOfPhases;
+        ++dwPhaseCount;
     }
-    pEngineState->userExperience.pUserExperience->OnApplyNumberOfPhases(dwNumberOfPhases);
+    pEngineState->userExperience.pUserExperience->OnApplyPhaseCount(dwPhaseCount);
 
     int nResult = pEngineState->userExperience.pUserExperience->OnApplyBegin();
     hr = UserExperienceInterpretResult(&pEngineState->userExperience, MB_OKCANCEL, nResult);