+++ src/libs/balutil/inc/BalBaseBootstrapperApplication.h
@@ -279,8 +279,8 @@ public: // IBurnUserExperience
     }
 
     // DEPRECATED: this will be merged with OnApplyBegin in wix4.
-    virtual STDMETHODIMP_(void) OnApplyNumberOfPhases(
-        __in DWORD /*dwNumberOfPhases*/
+    virtual STDMETHODIMP_(void) OnApplyPhaseCount(
+        __in DWORD /*dwPhaseCount*/
         )
     {
     }